Add text row to post card
