+++ feed/src/main/java/com/neaniesoft/warami/featurefeed/components/card/PostContentRow.kt
@@ -1,0 +1,29 @@
+package com.neaniesoft.warami.featurefeed.components.card
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.width
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import com.neaniesoft.warami.common.models.UriString
+
+@Composable
+fun PostContentRow(postTitle: String, thumbnailUrl: UriString?, url: UriString?) {
+    Row(
+        horizontalArrangement = Arrangement.SpaceBetween,
+        modifier = Modifier.padding(start = 16.dp, end = 16.dp, bottom = 16.dp),
+    ) {
+        Text(
+            text = postTitle,
+            style = MaterialTheme.typography.titleMedium,
+            modifier = Modifier.weight(1f),
+        )
+        Spacer(modifier = Modifier.width(16.dp))
+        PostContentRowThumbnail(thumbnailUrl = thumbnailUrl, url = url)
+    }
+}
