+++ ui/src/main/java/com/neaniesoft/warami/ui/nav/RootNavGraph.kt
@@ -1,6 +1,6 @@
 package com.neaniesoft.warami.ui.nav
 
-import com.neaniesoft.warami.featurefeed.FeaturefeedNavGraph
+import com.neaniesoft.warami.featurefeed.FeedNavGraph
 import com.neaniesoft.warami.signin.SigninNavGraph
 import com.neaniesoft.warami.ui.UiNavGraph
 import com.ramcosta.composedestinations.spec.DestinationSpec
@@ -13,7 +13,7 @@ object RootNavGraph : NavGraphSpec {
     override val startRoute: Route = UiNavGraph
     override val nestedNavGraphs: List<NavGraphSpec> = listOf(
         UiNavGraph,
-        FeaturefeedNavGraph,
+        FeedNavGraph,
         SigninNavGraph,
     )
 }
